requirements.txt
requests==2.27.1
numpy==1.22.2
pandas==1.4.0
scikit-learn==1.0.2
Pillow==7.0.0
matplotlib==3.5.1
streamlit==1.8.1